+++ changelog.md
@@ -2,7 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.2.0] - 2021-05-06 [LATEST]
+## [1.3.0] - 2021-05-09 [LATEST]
+
+### Added
+
+- Welcome/Leave Logs
+- Welcome/leave Message setup
+
+### Updated
+
+- Spam Messages UI
+- Delete Bot Messages Spam Feature
+- Year in Licince
+
+### Fixed
+
+- No Features error Bug in Serverinfo Command
+
+## [1.2.0] - 2021-05-06
 
 ### Added